+++ core/src/main/java/com/fizzed/blaze/core/Blaze.java
@@ -387,5 +387,4 @@ public class Blaze {
             });
         }
     }
-    
 }